Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    double a,b;
+    int a,b;
     scanf("%d%d",&a,&b);
     char c;
     scanf("%c",&c);
